requirements.txt
djangorestframework
djongo
django-cors-headers
pyjwt
Pillow
cloudinary